requirements.txt
langchain==0.0.333
fastapi==0.103.0
pinecone-client==2.2.4
openai==0.28.1
python-dotenv==1.0.0
uvicorn==0.23.2
pydantic-settings==2.0.3
pydantic==2.3.0
tiktoken==0.3.3
